改名

+++ kiri-engine/FileListen/HotReload.php
@@ -97,7 +97,7 @@ class HotReload extends Command
 	 */
 	public function healthCheck()
 	{
-		$pid = file_get_contents(storage('.swoole.pid'));
+		$pid = (int)file_get_contents(storage('.swoole.pid'));
 		$this->logger->debug('timer ticker.' . Process::kill($pid, 0));
 		if (!empty($pid) && !Process::kill($pid, 0)) {
 			if ($this->process && Process::kill($this->process->pid, 0)) {
